+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [unreleased]
+* update check for `disk_free_space()`
 
 #### 1.1.0 / 2021-09-01
 * automatically deactivate plugin after feature committed to core, currently set to `5.9-beta1`